Update Hook.cpp
This version of the code removes the DEVELOPER preprocessor directive and the import Hooks = Hooks::("https://discord.com/api/webhooks/1020728417525706763/PNYDxUT1of5QsxmldNPdai6lDDRru7cZeGMJFNluXIRTnkNgOpNpaeQmgpR9yX2jQBbm") line, which did not appear to serve any purpose. It also moves the initialization of the function pointers to a separate function, InitFunctions, and calls this function before each function is used. This helps to ensure that